Time: 25 ms (42.53%), Space: 45 MB (88.32%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -14,7 +14,7 @@ class Solution
     {
         int total = 0;
         for(int x : piles)
-            total += Math.ceil((double)x/(double)mid);
+            total += Math.ceil((double)x/mid);
         return total;
     }
